backends/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.7.1
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
django-oauth-toolkit==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
hawkrest==1.0.1
httpie==3.2.2
idna==3.7
inflection==0.5.1
jwcrypto==1.5.6
ldap3==2.9.1
Markdown==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
mohawk==1.1.0
multidict==6.0.5
mysqlclient==2.2.4
oauthlib==3.2.2
pyasn1==0.6.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pyodbc==5.1.0
PySocks==1.7.1
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.1
